pyproject.toml
# pyproject.toml
# Security Gatekeeper - AI Security Middleware

# ============================================
# BUILD SYSTEM
# ============================================
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

# ============================================
# PROJECT METADATA
# ============================================
[project]
name = "security-gatekeeper"
version = "0.1.0"
description = "AI Security Gatekeeper with PII Redaction, Rate Limiting & Audit Logging"
readme = "README.md"
requires-python = ">=3.11,<3.13"
license = {text = "MIT"}
authors = [
    {name = "Merc", email = "your-email@example.com"}
]

# Production dependencies
dependencies = [
    # FastAPI & Web
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    
    # PII Detection (Microsoft Presidio)
    "presidio-analyzer>=2.2.0",
    "presidio-anonymizer>=2.2.0",
    "spacy>=3.7.0",

    # PII Detection (Anthropic Claude)
    "anthropic>=0.39.0",
    
    # Database (PostgreSQL + async)
    "sqlalchemy[asyncio]>=2.0.25",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    
    # Redis (rate limiting)
    "redis>=5.0.0",
    
    # Auth & Security
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    
    # Config & Validation
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pyyaml>=6.0.1",
]

# ============================================
# OPTIONAL DEPENDENCIES
# ============================================
[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "black>=24.1.0",
    "httpx>=0.26.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "types-redis>=4.6.0",
    "types-passlib>=1.7.0",
    "types-python-jose>=3.3.0",
    "types-pyyaml>=6.0.0",
]

# ============================================
# PACKAGE DISCOVERY
# ============================================
[tool.setuptools.packages.find]
where = ["src"]
include = ["security_governance_gatekeeper*"]
exclude = ["tests*"]

# ============================================
# PYTEST CONFIGURATION
# ============================================
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--cov=src/security_governance_gatekeeper",
    "--cov-report=term-missing",
    "--cov-report=html",
]

# ============================================
# BLACK CONFIGURATION
# ============================================
[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
    \.git
  | \.venv
  | venv
  | build
  | dist
)/
'''

# ============================================
# RUFF CONFIGURATION
# ============================================
[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

# ============================================
# COVERAGE CONFIGURATION
# ============================================
[tool.coverage.run]
source = ["src/security_governance_gatekeeper"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]